feat(62) : 결재 반차 계산 추가, 커스텀 예외처리 (#63)
* refactor: 결재엔티티에 휴가요청 관련 위임 메서드 추가

* feat: 반차 로직 추가

* feat: 커스텀 예외 처리

* fix: 반차 수정

* style: 주석 제거

* fix: 테스트코드 임포트 오류 해결

* refactor: 안쓰는 메서드 제거

* refactor: 결재 엔티티 수정

* refactor: 결재 엔티티 수정

* style: todo 추가

+++ backend/team6/src/main/java/programmers/team6/domain/vacation/entity/VacationInfo.java
@@ -69,11 +69,11 @@ public class VacationInfo extends BaseEntity {
 		return this.version == version;
 	}
 
-	public void useVacation(int count) {
+	public void useVacation(double count) {
 		this.useCount += count;
 	}
 
-	public boolean canUseVacation(int count) {
+	public boolean canUseVacation(double count) {
 		return this.useCount + count <= totalCount;
 	}
